Make it compilable

+++ src/ScalingFitFCN.cxx
@@ -1,5 +1,5 @@
 #include "ScalingFitFCN.h"
-
+#include <cmath>
 
 double ScalingFitFCN::operator()(const double* par){
 
@@ -7,8 +7,8 @@ double ScalingFitFCN::operator()(const double* par){
 
 	double	scaling	= par[0];
 
-	for(size_t i=0;i<exptData.size();i++){
-		chisq	+= pow((scaling * calcData.at(i) - exptData.at(i))/exptUnc.at(i),2);
+	for(int i=0;i<exptData.size();i++){
+	  chisq	+= std::pow((scaling * calcData.at(i) - exptData.at(i))/exptUnc.at(i),2);
 	}
 
 	return	chisq;
